FIX: Don't corrupt users' last read IDs.
When we set up a chat channel, we subscribe to a message bus channel that notifies us about updates, one of them being new messages from other users. To show them in real-time, we add them immediately, but since we don't check if
all the latest ones are already loaded, we may create a gap where some messages will be missing until you refresh the tab.

We are changing how we handle new message updates and checking if it's possible to scroll into the future, and if it is, it means we haven't loaded all the latest messages yet, so we'll display a "jump to latest" button.

Thanks to this bug, we also noticed that our last read update logic is problematic because it aggressively updates to the latest unread and **loaded* message every two seconds without checking if the message is visible. Because of
this, the new message update that introduces the gap also updates your last read message to the latest channel's message.

Another scenario where this is problematic is when you visit a channel for one second, then switch to a different one. Here, the runner could mark up to 50 messages (max items loaded per page) as read. It also means the next time
you visit, we will start fetching from the unread message number 51, so you'll have to scroll up to find where you really left.

This PR implements a binary search to find an unread message available in the viewport and schedules a last read message update only when scrolling into the future, reducing the number of attempts to update. The downside is that we
end up with a few more requests per user, but the read progression feels more natural.