FIX: Don’t add users when chatable isn’t present
Currently the `AutoManageChannelMemberships` job will add users to a
channel following a set of rules. It won’t do anything if the channel
doesn’t exist or if the chatable isn’t a category.
But if the chatable doesn’t exist anymore then the job will raise an
exception.

This patch addresses this issue by simply checking if the chatable
exists before trying to do anything.