FIX: Auto-join users to automatic groups linked channels when granted staff status.
It depends on discourse/discourse#17663

The code to add or remove users from automatic groups uses mini_sql instead of `Group#add` or `Group#remove` as it needs to be able to work with a
large number of users. Because of this, when we grant staff status to a user or a TL promotion, it doesn't trigger the `user_added_to_group` event, and
we won't automatically join them to #staff linked channels. # Please enter the commit message for your changes. Lines starting