FIX: prevents mutation of dm creator selected users (#1100)
Before this fix, selecting a user which doesn’t have a DM yet, would result in the creation of a new channel using selected users as the list of users for the chatable. This would have the consequence of mutating each selected users in the dm creator screen and it would then be impossible to remove a user from the selected users array as the user given to `removeObject` would now be a different object than the one in the array.