FIX: Moving memberships out of channel decorator (#1131)
This commit makes a few key changes around how membership-related
data for chat channels is serialized, and cleans up the `ChatChannelFetcher`.

I have added a `current_user_membership` to the `ChatChannelSerializer`
which is used to store the current user's `UserChatChannelMembership`
data for that channel rather than storing that data on top level model attributes
for the channel. This has necessitated a lot of changes to the frontend and
tests and fixtures.

Also enabled by this change is the removal of the `previewing` property
in all the chat components -- instead we can now rely on the
`channel.current_user_membership.following` boolean which gets rid
of the bug where if there were more than N channels in the sidebar we
would not consider some of them followed even if the user was a member.

Finally I got rid of some deprecated usages of `current-user:main`, changing
to `service:current-user`, and same for the `session` service.