FIX: implements DeletedChatUser as a placeholder to deleted users (#1149)
This solution allows to implement all the logic server side and let the client side fully functional without any modification.

The only minor annoyance of the solution I know so far: deleted is translated server side which is not very useful, we might have to think of something different if that's prove to be an issue. However, system is not translated and it's fine so far.