FIX: Don't try to scroll to the last read message when fetching from latest.
Either clicking the jump arrow or sending a message when there are unloaded ones will fetch the latest page and scroll to the latest one. Calling
`_markLastReadMessage()` in these scenarios will break the scrolling since it also tries to scroll to the last read message immediately after, which we
can't guarantee it's on the loaded page.
 # Please enter the commit message for your changes. Lines starting