FIX: correctly links to channel message
Note this commit will require a rebake on affect posts to work.

This commit also fixed format-chat-date and copy message link codepaths to generate a correct URL to a chat message.

In the future it could be improved to simply rely on `channel.slug` both in frontend/backend but using `-` is  good solution for now as it will get replaced at runtime.