DEV: Reduce membership queries for DirectMessageChannelCreator
Previously, we created or updated UserChatChannelMembership
records in a loop, which added a lot of unnecessary SQL queries
when created a DM channel. For example, creating a new DM channel
for 4 users generated 12 queries (so N users * 3 queries):

* query to select the existing membership for the user
* query to select a channel from an ID (presumably AR doing
  this for the association on insert)
* query to insert the membership

Instead, we can do this in a single DB.exec call which does
2 queries regardless of how many users are used to create
the DM channel.

Relies on https://github.com/discourse/discourse-chat/pull/1179 for
tests to pass.