FEATURE: Allow changing DM channel notification settings (#1179)
This commit allows users to change the Mute settings
for DM channels as they can for regular category channels,
as well as changing the desktop and mobile notification
levels for the channel if it is not muted. It also changes
the unread_count logic for DM channels -- since the channels
can now be muted, the unread count does not always show regardless
for DM channels.