FIX: do not notify all users on dm channel creation
This commit makes a rather big and subtle change: creating a channel shouldn't create any kind of notification for anyone but the creator. Other users should receive notifications on the first message.

A side effect of this change is that a message on a direct message channel should always notify even if the user was not following the channel anymore, the way to go to avoid this, is to mute the channel.