REFACTOR: Trigger read updates through the visibility observer.
This change lets the visibility observer trigger a read update directly as soon as the message is visibile, simplifying the chat-live-pane, and making
the behavior easier to test.

The chat service handles the debouncing of the update action, and the chat-channel model, the ajax request and tracking of the last read message sent.

As a side note, I don't think remoinvg the document#querySelectorAll call it's possible because we need to recalculate the current last visible message
when the function fires since a scroll could have already happened.