FIX: Prevent deleting a category that has channels
Currently nothing prevents from deleting a category which has chat
enabled on it. This already led to a bug (see commit 07f9cb4).

This patch addresses this issue by not allowing the deletion of a
category if a channel is linked to it. This is the same approach we
already have for not deleting a category when topics or sub-categories
are present.