DEV: Make specs compatible with AR enums
Reviewable models from core are going to use `ActiveRecord` enums and
this change breaks one of our specs from this plugin.

This issue is addressed by using `#status_for_database` which returns
the value stored in the DB. This small change makes the specs compatible
with the current version of core and with the version that will use
`ActiveRecord` enums.