DEV: Refactor the ChatChannel model
Since its inception, the `ChatChannel` model evolved quite a bit.
Originally its `chatable` association allowed to link it to any existing
model but currently it’s only associated to two models: `Category` and
`DirectMessageChannel`.

Currently there are a lot of conditionals in the `ChatChannel` model to
determine logic based on the associated chatable. It’s a kind of “manual
polymorphism”.
This patch addresses this issue by making `ChatChannel` an STI model
which allows to create two new models: `CategoryChannel` and
`DMChannel`. Those models have dedicated logic allowing us to remove
the extra conditionals in the `ChatChannel` model. Then we only have one
entry point where we have some logic to choose which model to use based
on the provided chatable.