DEV: Refactor user_count calculation and chat membership management (#1251)
This commit intends to refactor and unify updates to chat channel
membership into a single `DiscourseChat::ChatChannelMembershipManager` class,
as well as making the `user_count` updates of the `ChatChannel` table
more consistent, as previously they happened in many places and it was
not always happening.

The `user_count` update is moved into a job, along with adding a new
`user_count_stale` column to the `ChatChannel` table. Whenever someone
is added/removed from the membership table, we update `user_count_stale`
to true _if it is not already_ and then enqueue a job in 3 seconds to update
the user count, which then sets `user_count_stale` back to false. This way
the query to count all the members, which can be expensive, is done out
of band, and the updates are throttled so we don't thrash the count unnecessarily.

We then send a MessageBus message to the client about the channel
metadata change, and if the user is on the member list for the channel
we update the count and reload the list of members.