FEATURE: Improve chat messages flagging.
This PR re-uses core's flag modal so users can select a flag type and send PMs when selecting notify_user or notify_moderators. It also lets staff take actions, or queue chat messages for review.

One difference is that we don't have the concept of "hiding" as we do for posts, so we have to delete messages instead on those scenarios.