FIX: enforces chat_channel_id when present
This will deserve an heavier refactoring in the future, and more importantly moving all this logic under something like `/chat/chat_channels/:id/messages?target_message_id=865` to enforce message being a child from the channel.

This commit however prevents the issue for now and ensures this case is tracked for future refactoring with a test.