UX: moves new channel button to browse view (#1278)
Removes the dropdown with browse/create from the sidebar. As a side effect:

it's now a button using the pencil icon, leading to browse page
the new channel button is at the top of browse page
It's only changing behavior for staff, regular users were already not seeing the dropdown as they can’t create channels.