FIX: direct_message_enabled_groups refinements (#1285)
Users who are not inside any direct_message_enabled_groups
should not be able to create DM channels, however if they
are messaged by a user who does have this permission they
should be able to send messages back in the channel without
issue. Update setting description to reflect this.

Also do not show the create direct message channel button
in the sidebar for users who do not have permission to do
so, where the sidebar is showing for them since another user
has opened a DM channel with them.