FIX: Remove hijack on webhook and improve documentation
This commit removes the `hijack` on the chat webhooks controller
since it is of dubious necessity and because `hijack` is
suppressing the correct resposne codes.

Also add some documentation around how the incoming webhooks
work, their params, and their limits.

Bump the incoming text param length limit to 2000, 1000 is
quite low.