FIX: Don't fail when trying to display DM flags in the review queue. (#1323)
The serializer tried to serialize the whole message, but this needed the channel passed in the options hash to work. I wrote the code this way to avoid fetching each message chat channel when serializing in the context of the `ChatView`.

This change simplifies the reviewable chat message serializer and fallbacks to the object's `chat_channel` when `@options[:chat_channel]` is `nil`.