FIX: improves iOS hack for momentum scrolling
safari iOS has the bug when programmatically scrolling to a part of the scrollable element while being in momentum scrolling. This will cause the renderer to lag and cause an empty screen. In the past `-webkit-overflow-scrolling: touch` was used to stop momentum but it's not longer usable since iOS 13. This commit attempts to implement a decent solution to the problem.